Separate models from adapter code
